+++ infer/src/base/CommandDoc.ml
@@ -215,10 +215,11 @@ $(b,infer) $(i,[options])|}
              as argument. The format is one option per line, and enclosing single ' and double \" \
              quotes are ignored."
         ; `P
-            "Options without a default value (e.g., $(b,--linter)) and options with list-like \
-             values (e.g., $(b,--Xbuck)) all have a corresponding $(b,--option-reset) flag that \
-             resets their values to nothing or the empty list, respectively. For instance, \
-             $(b,--Xbuck-reset) will cancel any previous $(b,--Xbuck) option passed to infer."
+            "Options without a default value (e.g., $(b,--changed-files-index)) and options with \
+             list-like values (e.g., $(b,--Xbuck)) all have a corresponding $(b,--option-reset) \
+             flag that resets their values to nothing or the empty list, respectively. For \
+             instance, $(b,--Xbuck-reset) will cancel any previous $(b,--Xbuck) option passed to \
+             infer."
         ; `P
             "See the manuals of individual infer commands for details about their supported \
              options. The following is a list of all the supported options (see also \