+++ infer/src/unit/DifferentialTestsUtils.ml
@@ -10,8 +10,8 @@ open! IStd
 let create_fake_jsonbug ?(bug_type = "bug_type") ?(qualifier = "qualifier") ?(severity = "severity")
     ?(line = 1) ?(column = 1) ?(procedure = "procedure") ?(procedure_start_line = 1)
     ?(file = "file/at/a/certain/path.java") ?(bug_trace = []) ?(key = "File|method|TYPE")
-    ?(node_key = Some "1234") ?(hash = "1") ?(dotty = None) ?(infer_source_loc = None)
-    ?(linters_def_file = Some "file/at/certain/path.al") ?doc_url () : Jsonbug_t.jsonbug =
+    ?(node_key = Some "1234") ?(hash = "1") ?(dotty = None) ?(infer_source_loc = None) () :
+    Jsonbug_t.jsonbug =
   { bug_type
   ; qualifier
   ; severity
@@ -27,8 +27,6 @@ let create_fake_jsonbug ?(bug_type = "bug_type") ?(qualifier = "qualifier") ?(se
   ; dotty
   ; infer_source_loc
   ; bug_type_hum= bug_type
-  ; linters_def_file
-  ; doc_url
   ; traceview_id= None
   ; censored_reason= None
   ; access= None